pyproject.toml
[project]
name = "restaurant-sentiment"
version = "0.1.0"
description = "Dynamic restaurant sentiment analysis using web scraping and AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.24.0",
    "jinja2>=3.1.6",
    "pandas>=2.3.0",
    "tavily-python>=0.7.9",
]
